rgw: fix signed char truncation in delimiter check

+++ src/rgw/rgw_rados.cc
@@ -2435,7 +2435,7 @@ static inline std::string after_delim(std::string_view delim)
   char e = delim.back();
   delim.remove_suffix(1);
   std::string result{delim.data(), delim.length()};
-  if (e < 255) {
+  if (uint8_t(e) < 255) {
     result += char(++e);
   } else {
     result += e;
